Blood Bank & Donor Management System
Blood Bank & Donor Management System is a web based system aimed at maintaining all the information related to blood donors, different blood groups available in each blood bank and help them manage in a better way. It helps to record the various people that offer to donate blood to the people who need it in the hospitals and donation centers. It will be designed to serve all people and reduce on the manual recordings of people in books which brings confusion and alteration of records.
It will be composed of the administration who will be in charge of managing the user input, deletion of invalid details, updating and alteration of the records among others into the system.
User Feature’s (Front end):
•	Responsive Template, Mobile Friendly
•	Easy to use
•	Blood Donor Registration System
•	Blood Group and location-based search engine
•	Donor Details Information
•	Contact us Inquiry

Admin Feature’s (Administration Section):
•	Manage Blood Group(add, delete)
•	Add and Delete Donor Information
•	Enable and Disable Donors
•	Manage Website Pages
•	Manage to Contact us Queries
•	Update the contact us Info
•	Admin Dashboard